requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
distro==1.8.0
docx==0.2.4
effdet==0.4.1
emoji==2.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.46.0
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
iopath==0.1.10
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
JPype1==1.4.1
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-client==8.6.0
jupyter-core==5.5.0
kiwisolver==1.4.5
langchain==0.0.340
langdetect==1.0.9
langsmith==0.0.71
layoutparser==0.3.4
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==0.28.0
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
pikepdf==8.10.1
Pillow==10.1.0
platformdirs==4.1.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyarrow==14.0.1
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.2
pydantic-core==2.14.5
pydeck==0.8.1b0
pygments==2.17.2
pyparsing==3.1.1
pypdf==3.17.2
pypdf2==3.0.1
pypdfium2==4.25.0
pyreadline3==3.4.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-iso639==2023.12.11
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.5.2
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
safetensors==0.4.1
scipy==1.11.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
streamlit==1.28.1
sympy==1.12
tabula-py==2.9.0
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
timm==0.9.12
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.36.1
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
unstructured==0.11.4
unstructured-client==0.15.0
unstructured-inference==0.7.18
unstructured.pytesseract==0.3.12
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.12
wrapt==1.16.0
xlrd==2.0.1
yarl==1.9.4
zipp==3.17.0
